bug fix, not finished yet. One last bug
